pyproject.toml
[project]
name = "illyrion"
version = "0.1.1"
description = " Discord Bot for Personal use"
requires-python = ">=3.13"
dependencies = [
    "py-cord",
    "audioop-lts",
    "psutil",
    "python-dotenv>=1.1.1",
    "black>=25.1.0",
]

authors = [{ name = "Xodium", email = "info@xodium.org" }]
license = "AGPL-3.0"
readme = "README.md"
keywords = ["discord", "bot", "personal"]
